requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.3
uvicorn>=0.15.0,<0.16.0
settrade-v2==2.1.4
python-decouple==3.8